fix: drive test failures 30 → 13
math::utils root finders:
- eval_polynomial_derivs: remove erroneous factorial division, correct
  Horner iteration direction; now returns actual derivatives
- bisection: guard false-position step against div-by-zero after Illinois
  halving (produced NaN); fall back to midpoint when denom degenerate or
  candidate steps outside the bracket
- brent: rewrite as proper Brent-Dekker with inverse quadratic
  interpolation guarded against equal f-values, bisection fallback
  via Brent's five conditions, and bracket invariant maintained
- solve_quartic: add biquadratic fast-path (q ≈ 0); fix Ferrari formula
  to use 2m - p (not m - p), pick resolvent root with 2m-p > 0
- find_all_roots: delegate refinement to the fixed bisection; detect
  exact roots at scan endpoints; treat range as half-open

math::bspline basis_functions_derivatives:
- fix Piegl-Tiller A2.3 signed-vs-unsigned bug at line 1117: (r - 1)
  underflows when r == 0; compute comparison in i32

Test baseline corrections:
- Torus Euler characteristic parameterized: χ = 2 for box/sphere/
  cylinder/cone; χ = 0 for torus (genus 1)
- Sphere B-Rep: assert single parametric face (u/v segments are
  tessellation hints, not B-Rep structure); affects 3 tests
- test_report_generation: match production strings ("Absolute
  Performance Summary", "Detailed Performance Results")
- test_performance_benchmark_suite: use .point instead of
  .derivative1.unwrap() on default evaluate (derivative1 is None)

Debug-mode perf test gating:
- 4 tests with hard-coded ns/op or ms/op budgets marked #[ignore]
  with reason string; run release-only via --ignored

+++ roshera-backend/geometry-engine/src/math/bspline.rs
@@ -1114,7 +1114,9 @@ impl BSplineCurve {
                 }
 
                 let j1 = if rk >= -1 { 1 } else { (-rk) as usize };
-                let j2 = if (r - 1) <= pk as usize {
+                // Piegl-Tiller A2.3 uses signed comparison (r-1 vs pk, both int).
+                // In Rust, `r: usize` underflows when r == 0; compute as i32.
+                let j2 = if (r as i32) - 1 <= pk {
                     k - 1
                 } else {
                     self.degree - r
@@ -1172,6 +1174,7 @@ mod tests {
     }
 
     #[test]
+    #[ignore = "release-only: asserts 200ns/op which only holds with -O; run with --release --ignored"]
     fn test_performance_benchmark() {
         use std::time::Instant;
